requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.24.0
openpyxl>=3.1.0
fpdf2>=2.7.0
